// src/core/models/IThreeDQueryResponse.ts
export interface IThreeDQueryResponse {
  acquirertransactionreference: string;
  acsurl: string;
  enrolled: string;
  threedpayload: string;
  transactionreference: string;
  transactionstartedtimestamp?: string;
  customeroutput?: string;
  livestatus?: string;
  merchantname?: string;
  dccenabled?: string;
  settleduedate?: string;
  errorcode?: string;
  tid?: string;
  merchantnumber?: string;
  merchantcountryiso2a?: string;
  threedversion?: string;
  paymenttypedescription?: string;
  accounttypedescription?: string;
  acquirerresponsecode?: string;
  requesttypedescription?: string;
  issuer?: string;
  maskedpan?: string;
  errormessage?: string;
  operatorname?: string;
  issuercountryiso2a?: string;
  settlestatus?: string;
}